apps/backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
distro==1.9.0
fastapi==0.133.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.24.0
packaging==26.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pgvector==0.4.1
numpy>=1.26.0
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
wrapt==2.1.1
PyJWT==2.10.1
email-validator==2.2.0
sentry-sdk[fastapi]==2.29.1

